+++ slam-libraries/viam-cartographer/src/slam_service/slam_service.h
@@ -261,7 +261,7 @@ class SLAMServiceImpl final : public SLAMService::Service {
     // SetUpSLAM sets the correct action mode, prepares the map builder and
     // loads the right hyperparameters based on the action mode. Needs to be
     // called before running slam.
-    int SetUpSLAM();
+    double SetUpSLAM();
 
     // GetLatestPaintedMapSlices paints and returns the current map of
     // Cartographer
